requirements.txt
base64
codecs
pycryptodome